preprocessing
This notebook is dedicated to the preprocessing of the Simple Gender Classification dataset, which contains demographic data such as age, height, weight, occupation, and more. The goal of this notebook is to clean and transform the data, including handling missing values, encoding categorical features, and scaling numerical features for further use in machine learning models.